Rename module, package and CLI to lls-*
