+++ plugin/napari_lattice/napari.yaml
