+++ plugin/napari_lattice/ui_core.py
@@ -8,9 +8,9 @@ from aicsimageio import AICSImage
 import pyclesperanto_prototype as cle
 from .utils import pad_image_nearest_multiple, check_dimensions
 
-from lattice_lightsheet_core.llsz_core import pycuda_decon, skimage_decon
-from lattice_lightsheet_core import config, DeskewDirection, DeconvolutionChoice, SaveFileType
-from lattice_lightsheet_core .io import save_img
+from lls_core.llsz_core import pycuda_decon, skimage_decon
+from lls_core import config, DeskewDirection, DeconvolutionChoice, SaveFileType
+from lls_core .io import save_img
 
 if TYPE_CHECKING:
     from napari.types import ImageData
